Check a meteorology data file for compliance with the PEcAn standard

Check a meteorology data file for compliance with the PEcAn standard

check_met_input_file(metfile, variable_table = pecan_standard_met_table,
  required_vars = variable_table %>% dplyr::filter(is_required) %>%
  dplyr::pull(cf_standard_name), warn_unknown = TRUE)

Arguments

metfile
Path of met file to check, as a scalar character.

variable_table
`data.frame` linking standard names to their units. Must contain columns "cf_standard_name" and "units". Default is [pecan_standard_met_table].

required_vars
Character vector of required variables. Defaults to variables marked as required in `variable_table`.

warn_unknown
Logical. If `TRUE` (default), throw a warning for variables not in `variable_table`. Otherwise, ignore unknown variables.

Value

`data.frame` summarizing the results of the tests.