Download ERA 5 data

Link to full data documentation.

download.ERA5.old(
  outfolder,
  start_date,
  end_date,
  lat.in,
  lon.in,
  product_types = "all",
  overwrite = FALSE,
  reticulate_python = NULL,
  ...
)

Arguments

outfolder

Directory where results should be written

start_date, end_date

Range of years to retrieve. Format is YYYY-MM-DD.

lat.in, lon.in

Site coordinates, decimal degrees (numeric)

product_types

Character vector of product types, or "all". Must be one or more of: "reanalysis", "ensemble members", "ensemble mean", "ensemble spread"

overwrite

Logical. If FALSE (default), skip any files with the same target name (i.e. same variable) that already exist in outfolder. If TRUE, silently overwrite existing files.

reticulate_python

Path to Python binary for reticulate (passed to reticulate::use_python()). If NULL (default), use the system default.

...

Currently unused. Allows soaking up additional arguments to other methods.

Value

Character vector of file names containing raw, downloaded data (invisibly)

Details

Under the hood, this function uses the Python cdsapi module, which can be installed via pip (pip install --user cdsapi). The module is accessed via the reticulate package.

Using the CDS API requires you to create a free account at https://cds.climate.copernicus.eu. Once you have done that, you will need to configure the CDS API on your local machine by creating a ${HOME}/.cdsapi file, as described here.

Author

Alexey Shiklomanov

Examples

if (FALSE) {
files <- download.ERA5(
  "ERA5_output",
  start_date = "2010-01-01",
  end_date = "2010-02-01",
  lat.in = 45.5594,
  lon.in = -84.6738,
  product_types = "all"
)
}