download.Fluxnet2015

Download Fluxnet 2015 CSV files

Usage

download.Fluxnet2015(
  sitename,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  verbose = FALSE,
  username = "pecan",
  ...
)

Arguments

sitename

the FLUXNET ID of the site to be downloaded, used as file name prefix. The 'SITE_ID' field in list of Ameriflux sites

outfolder

location on disk where outputs will be stored

start_date

the start date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year part of the date)

end_date

the end date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year part of the date)

overwrite

should existing files be overwritten

verbose

should the function be very verbose

username

login name for Ameriflux

...

further arguments, currently ignored

Author

Ankur Desai, based on download.Ameriflux.R by Josh Mantooth, Rob Kooper