Download ICOS Ecosystem data products

Currently available products: Drought-2018 ecosystem eddy covariance flux product https://www.icos-cp.eu/data-products/YVR0-4898 ICOS Final Fully Quality Controlled Observational Data (Level 2) https://www.icos-cp.eu/data-products/ecosystem-release

download.ICOS(
  sitename,
  outfolder,
  start_date,
  end_date,
  product,
  overwrite = FALSE,
  ...
)

Arguments

sitename

ICOS id of the site. Example - "BE-Bra"

outfolder

path to the directory where the output file is stored. If specified directory does not exists, it is created.

start_date

start date of the data request in the form YYYY-MM-DD

end_date

end date area of the data request in the form YYYY-MM-DD

product

ICOS product to be downloaded. Currently supported options: "Drought2018", "ETC"

overwrite

should existing files be overwritten. Default False.

...

used when extra arguments are present.

Value

information about the output file

Author

Ayush Prasad

Examples

if (FALSE) {
download.ICOS("FI-Sii", "/home/carya/pecan", "2016-01-01", "2018-01-01", product="Drought2018")
}