Download NEON Site Met CSV files

download.NEONmet

download.NEONmet(
  sitename,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)

Arguments

sitename

the NEON ID of the site to be downloaded, used as file name prefix. The 4-letter SITE code in list of NEON sites

outfolder

location on disk where outputs will be stored

start_date

the start date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year and month of the date)

end_date

the end date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year and month part of the date)

overwrite

should existing files be overwritten

verbose

makes the function output more text

...

further arguments, currently ignored

Details

Uses NEON v0 API to download met data from NEON towers and convert to CF NetCDF

Examples

if (FALSE) {
result <- download.NEONmet('HARV','~/','2017-01-01','2017-01-31',overwrite=TRUE)
}