download.raw.met.module

download.raw.met.module

.download.raw.met.module(
  dir,
  met,
  register,
  machine,
  start_date,
  end_date,
  str_ns,
  con,
  input_met,
  site.id,
  lat.in,
  lon.in,
  host,
  site,
  username,
  overwrite = FALSE,
  dbparms,
  Ens.Flag = FALSE
)

Arguments

dir

directory to write outputs to

met

source included in input_met

register

register.xml, provided by met.process

machine

machine associated with hostname, provided by met.process

start_date

the start date of the data to be downloaded (will only use the year part of the date)

end_date

the end date of the data to be downloaded (will only use the year part of the date)

str_ns

substitute for site_id if not provided, provided by met.process

con

database connection based on dbparms in met.process

input_met

Which data source to process

site.id

site id

lat.in

site latitude, provided by met.process

lon.in

site longitude, provided by met.process

host

host info from settings file

site

site info from settings file

username

database username

overwrite

whether to force download.raw.met.module to proceed

dbparms

database settings from settings file

Ens.Flag

default set to FALSE

Value

A list of data frames is returned containing information about the data file that can be used to locate it later. Each data frame contains information about one file.