model.train

Function to create linear regression models for specific met variables. This is used in conjunction with temporal.downscale.functions() to generate linear regression statistics and save their output to be called later in lm_ensemble_sims().

Usage

model.train(dat.subset, v, n.beta, resids = resids, threshold = NULL, ...)

Arguments

dat.subset

data.frame containing lags, next, and downscale period data

v

variable name, as character

n.beta

number of betas to pull from

resids

TRUE or FALSE, whether to use residuals or not

threshold

NULL except for surface_downwelling_shortwave_radiation, helps with our distinction between day and night (no shortwave without sunlight)

...

further arguments, currently ignored

Details

TDM Model Train Linear regression calculations for specific met variables

See also

Author

Christy Rollinson, James Simkins