Extract and temporally downscale points from downloaded grid files

Extract and temporally downscale points from downloaded grid files

Usage

process_gridded_noaa_download(
  lat_list,
  lon_list,
  site_id,
  downscale,
  overwrite,
  forecast_date,
  forecast_time,
  model_name,
  model_name_ds,
  model_name_raw,
  output_directory
)

Arguments

lat_list

lat for site

lon_list

lon for site

site_id

Unique site_id for file creation

downscale

Logical. Default is TRUE. Downscales from 6hr to hourly

overwrite

Logical. Default is FALSE. Should exisiting files be overwritten

forecast_date

Date for download

forecast_time

Time (0,6,12,18) for start of download

model_name

Name of model for file name

model_name_ds

Name of downscale file name

model_name_raw

Name of raw file name

output_directory

Output directory

Value

List