save.betas

Function to save betas as a .nc file. This is utilized in gen.subdaily.models() when linear regression models are created

Usage

save.betas(model.out, betas, outfile)

Arguments

model.out

list linear regression model output

betas

name of the layer of betas to save (e.g. 'betas' or 'betas.resid')

outfile

location where output will be stored

Details

TDM Save Betas Saves betas that are calculated during gen.subdaily.models()

See also

Author

Christy Rollinson, James Simkins