Download Ameriflux LBL CSV files

download.AmerifluxLBL. Function uses amf_download_base function from amerifluxr package to download a zip-file of data. The zip-file is extracted to a csv-file that is stored to the given outfolder. Details about amf_download_base function can be found here: https://github.com/chuhousen/amerifluxr/blob/master/R/amf_download_base.R

Usage

download.AmerifluxLBL(
  sitename,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  verbose = FALSE,
  username = "pecan",
  method,
  useremail = "@",
  data_product = "BASE-BADM",
  data_policy = "CCBY4.0",
  ...
)

Arguments

sitename

the Ameriflux ID of the site to be downloaded, used as file name prefix. The 'SITE_ID' field in list of Ameriflux sites

outfolder

location on disk where outputs will be stored

start_date

the start date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year part of the date)

end_date

the end date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year part of the date)

overwrite

should existing files be overwritten

verbose

should the function be very verbose

username

Ameriflux username

method

Optional. download_file() function option. Use this to set custom programs such as ncftp

useremail

Used email, should include 'address sign' for code to be functional

data_product

AmeriFlux data product

data_policy

Two possible licenses (based on the site): 'CCBY4.0' or 'LEGACY'

...

further arguments, currently ignored

Details

Uses Ameriflux LBL JSON API to download met data from Ameriflux towers in CSV format

Author

Ankur Desai, Henri Kajasilta based on download.Ameriflux.R by Josh Mantooth, Rob Kooper, Shawn Serbin

Examples

if (FALSE) { # \dontrun{
options(download.ftp.method = "ncftpget")
result <- download.AmerifluxLBL("US-Akn","~/","2011-01-01","2011-12-31",overwrite=TRUE)
} # }