Download NOAA GEFS Weather Data

Download NOAA GEFS Weather Data

Usage

download.NOAA_GEFS(
  site_id,
  lat.in,
  lon.in,
  outfolder,
  start_date = Sys.Date(),
  end_date = start_date + lubridate::days(16),
  downscale = TRUE,
  overwrite = FALSE,
  ...
)

Arguments

site_id

The unique ID given to each site. This is used as part of the file name.

lat.in

site latitude in decimal degrees

lon.in

site longitude in decimal degrees

outfolder

Directory where results should be written

start_date

Range of dates/times to be downloaded (default assumed to be time that function is run)

end_date

end date for range of dates to be downloaded (default 16 days from start_date)

downscale

logical, assumed True. Indicates whether data should be downscaled to hourly

overwrite

logical. Download a fresh version even if a local file with the same name already exists?

...

Additional optional parameters, currently ignored

Value

A list of data frames is returned containing information about the data file that can be used to locate it later. Each data frame contains information about one file.

Information on Units

Information on NOAA weather units can be found below. Note that the temperature is measured in degrees C, but is converted at the station and downloaded in Kelvin.

NOAA_GEFS General Information

This function downloads NOAA GEFS weather data. GEFS is an ensemble of 31 different weather forecast models. A 16 day forecast is available every 6 hours and a 35 day forecast is available every 24 hours. Both are at 3-hour frequency for the first 10 days of the forecast and 6-hour frequency beyond that. Each forecast includes information on a total of 8 variables. These are transformed from the NOAA standard to the internal PEcAn standard.

Data Availability

NOAA GEFS weather data is available on a rolling 4 day basis. Dates provided in "start_date" must be within this range. The end date can be any point after that, but if the end date is beyond 16 days (35 days for the midnight UTC forecast), only 16 (35) days worth of forecast are retrieved. Times are rounded down to the previous 6 hour forecast.

NOAA GEFS weather data isn't always posted immediately. Each 16-day forecast takes approximately three hours to run, and the once-a-day forecasts for days 17-35 are posted much later (up to 21 hours) than the forecasts for days 0 to 16. See the [GEFS v12 release announcement](https://www.weather.gov/media/notification/pdf2/scn20-75gefs_v12_changes.pdf) for details.

Data Save Format

Data is saved in the netcdf format to the specified directory. File names reflect the precision of the data to the given range of days. NOAA.GEFS.willow creek.3.2018-06-08T06:00.2018-06-24T06:00.nc specifies the forecast, using ensemble number 3 at willow creek on June 6th, 2018 at 6:00 a.m. to June 24th, 2018 at 6:00 a.m.

References

https://www.ncdc.noaa.gov/crn/measurements.html

Author

Quinn Thomas, modified by K Zarada

Examples

if (FALSE) { # \dontrun{
 download.NOAA_GEFS(outfolder="~/Working/results",
    lat.in= 45.805925,
    lon.in = -90.07961,
    site_id = 676)
} # }