temporal_downscale_functions

This function contains the functions that do the heavy lifting in gen.subdaily.models() and predict.subdaily.workflow(). Individual variable functions actually generate the models and betas from the dat.train_file and save them in the output file. save.model() and save.betas() are helper functions that save the linear regression model output to a specific location. In the future, we should only save the data that we actually use from the linear regression model because this is a large file. predict.met() is called from predict.subdaily.workflow() and references the linear regression model output to predict the ensemble data.

temporal.downscale.functions(dat.train, n.beta, day.window,
  resids = FALSE, parallel = FALSE, n.cores = NULL,
  seed = format(Sys.time(), "%m%d"), outfolder,
  print.progress = FALSE, ...)

Arguments

dat.train
- training data generated by tdm_nc2dat.train.R

n.beta
- number of betas to generate

day.window
- number of days surrounding current day we want to pull statistics from

resids
- whether or not to propogate residuals, set to FALSE

parallel
- whether or not to run in parallel. this is a feature still being worked on, set to FALSE

n.cores
- number of cores to use parallel processing on, set to NULL

seed
- allows this to be reproducible

outfolder
= where the output should be stored

print.progress
- print progress of model generation?

Details

Temporal Downscale Functions Met variable functions that are called in gen.subdaily.models and predict.subdaily.workflow

See also